authentication

+++ server/app/routers/cameras.py
@@ -1,10 +1,12 @@
+from app.auth import BaseAuth, check_is_admin, get_auth, get_user
 from app.controllers.cameras import CameraController
 from app.database import get_session
 from app.schemas.cameras import Camera, CameraCreate, CameraUpdate
-from fastapi import APIRouter, Depends, HTTPException
+from app.schemas.users import User
+from fastapi import APIRouter, Depends, HTTPException, status, Response
 from fastapi.responses import StreamingResponse
 from sqlalchemy.orm import Session
-from typing import List
+from typing import List, Optional
 from uuid import UUID
 
 
@@ -13,6 +15,7 @@ router = APIRouter()
 
 @router.get('/', response_model=List[Camera])
 async def get_cameras(
+    user: User = Depends(get_user),
     db: Session = Depends(get_session)
 ) -> List[Camera]:
     """
@@ -24,22 +27,26 @@ async def get_cameras(
 @router.post('/', response_model=Camera)
 async def create_camera(
     payload: CameraCreate,
+    user: User = Depends(get_user),
     db: Session = Depends(get_session)
 ) -> Camera:
     """
     Create camera
     """
+    check_is_admin(user)
     return CameraController.create_camera(db, payload)
 
 
 @router.get('/{camera_id}', response_model=Camera)
 async def get_camera(
     camera_id: UUID,
+    user: User = Depends(get_user),
     db: Session = Depends(get_session)
 ) -> Camera:
     """
     Get camera
     """
+    check_is_admin(user)
     return CameraController.get_camera(db, camera_id)
 
 
@@ -47,33 +54,41 @@ async def get_camera(
 async def update_camera(
     camera_id: UUID,
     payload: CameraUpdate,
+    user: User = Depends(get_user),
     db: Session = Depends(get_session)
 ) -> Camera:
     """
     Update camera
     """
+    check_is_admin(user)
     return CameraController.update_camera(db, camera_id, payload)
 
 
 @router.delete('/{camera_id}')
 async def delete_camera(
     camera_id: UUID,
+    user: User = Depends(get_user),
     db: Session = Depends(get_session)
 ) -> None:
     """
     Delete camera
     """
+    check_is_admin(user)
     return CameraController.delete_camera(db, camera_id)
 
 
 @router.get('/{camera_id}/live')
 async def stream_camera(
     camera_id: UUID,
+    token: Optional[str] = None,
+    auth: BaseAuth = Depends(get_auth),
     db: Session = Depends(get_session)
 ):
     """
     Get camera stream
     """
+    user = auth.get_user_with_token(token, db)
+
     try:
         stream = CameraController.get_stream(db, camera_id)
