+++ server/app/routers/identities.py
@@ -1,8 +1,10 @@
+from app.auth import check_is_admin, get_user
 from app.controllers.identities import IdentityController
 from app.controllers.recognition import RecognitionController
 from app.database import get_session
 from app.schemas.identities import Identity, IdentityCreate, IdentityUpdate
 from app.schemas.recognition import FaceEncoding
+from app.schemas.users import User
 from fastapi import APIRouter, Depends, File, HTTPException, UploadFile
 from sqlalchemy.orm import Session
 from typing import List
@@ -14,6 +16,7 @@ router = APIRouter()
 
 @router.get('/', response_model=List[Identity])
 async def get_identities(
+    user: User = Depends(get_user),
     db: Session = Depends(get_session)
 ) -> List[Identity]:
     """
@@ -25,17 +28,20 @@ async def get_identities(
 @router.post('/', response_model=Identity)
 async def create_identity(
     payload: IdentityCreate,
+    user: User = Depends(get_user),
     db: Session = Depends(get_session)
 ) -> Identity:
     """
     Create identity
     """
+    check_is_admin(user)
     return IdentityController.create_identity(db, payload)
 
 
 @router.get('/{identity_id}', response_model=Identity)
 async def get_identity(
     identity_id: UUID,
+    user: User = Depends(get_user),
     db: Session = Depends(get_session)
 ) -> Identity:
     """
@@ -48,22 +54,26 @@ async def get_identity(
 async def update_identity(
     identity_id: UUID,
     payload: IdentityUpdate,
+    user: User = Depends(get_user),
     db: Session = Depends(get_session)
 ) -> Identity:
     """
     Update identity
     """
+    check_is_admin(user)
     return IdentityController.update_identity(db, identity_id, payload)
 
 
 @router.delete('/{identity_id}')
 async def delete_identity(
     identity_id: UUID,
+    user: User = Depends(get_user),
     db: Session = Depends(get_session)
 ) -> Identity:
     """
     Delete identity
     """
+    check_is_admin(user)
     return IdentityController.delete_identity(db, identity_id)
 
 
@@ -71,11 +81,14 @@ async def delete_identity(
 async def learn(
     identity_id: UUID,
     picture: UploadFile = File(...),
+    user: User = Depends(get_user),
     db: Session = Depends(get_session)
 ) -> FaceEncoding:
     """
     Learn face on a picture
     """
+    check_is_admin(user)
+
     identity = IdentityController.get_identity(db, identity_id)
     image = RecognitionController.load_uploaded_file(picture)
     faces = RecognitionController.get_faces_locations(image)
